LLM/workspace/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
opencv-python>=4.8.0
pillow>=10.0.0


